+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## [1.0.0+69]
+## [1.0.0+74]
 
 - feat: Enhance NotificationsController with category parameters and update version to 1.0.0+74